+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - fix: tooltip behavior with shouldBegin problem, closes #2016;
 - fix: the position of grid plugins when there is something on the top of the canvas, closes: #2012;
 - fix: fisheye destroy and new problem, closes: #2018;
+- fix: node event with wrong canvasX and canvasY problem, closes: #2027;
 - feat: improve the performance on the combos;
 - feat: hide shapes beside keyShape while zooming.